backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis[hiredis]==5.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.4.0
aiohttp==3.10.10
aiosmtplib==3.0.2
cryptography==43.0.3
scikit-learn==1.5.2
numpy==2.1.3
scipy==1.14.1
psutil==6.1.0
reportlab==4.2.5
httpx==0.27.2
python-multipart==0.0.12
anyio==4.6.2
tenacity==9.0.0
